+++ src/pages/App.tsx
@@ -6,7 +6,7 @@ import AddressClaimModal from '../components/claim/AddressClaimModal'
 import Header from 'components/Header'
 import Polling from '../components/Header/Polling'
 import Popups from 'components/Popups'
-import Web3ReactManager from '../custom/components/Web3ReactManager'
+import Web3ReactManager from '../components/Web3ReactManager'
 import ErrorBoundary from '../components/ErrorBoundary'
 import { ApplicationModal } from '../state/application/actions'
 import { useModalOpen, useToggleModal } from '../state/application/hooks'
